requirements.txt
backoff==1.10.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
opencage==1.2.1
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
ratelim==0.1.6
redis==3.5.3
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
